requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
google-genai
pillow